pyproject.toml
[project]
name = "physics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "ipykernel>=7.2.0",
    "numpy>=2.4.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
